Subject: Merge rev 102 from s48/tuebingen/syslog [s48-tuebingen @ 103]
This fixes a regression, but I'm not sure the fix is The Right Thing:

After a callback, the stack frame of the callback didn't get cleared.
This meant that one callback works but two in a row fail.  So I'm just
removing the current frame at the end of S48-RESTART.