Subject: Fix hygiene bug in DEFINE-RECORD-TYPE [s48-tuebingen @ 120]
DEFINE-RECORD-TYPE would expand into the wrong thing if there was a
field name identical to the name of the type that appears as a
constructor argument; the constructor used that same name as a
parameter, shadowing the name for the type.  The fix was to rename
those parameters before including them in the output.