Subject: Make SPAWN return the created thread [s48-tuebingen @ 363]
Make SPAWN return the thread object associated with the spawned
thread.  This required some restructuring of the code, as the thread
object used to be created in the scheduler.  It's now being created in
the thread from which SPAWN was called with empty slots for the
scheduler and the dynamic environment; these are filled in by the
event handler for SPAWNED.

Also export CURRENT-THREAD from THREADS instead from THREADS-INTERNAL.