Subject: Store offset to template in continuation data [s48-tuebingen @ 616]
The native code needs to know the template to restart the VM. It
therefore stores the template in a register. Loading this register
when invoking a continuation requires knowledge of the location of the
template within the stack. This patch stores this offset in the
arguments of the CONT-DATA pseudo instruction with #xffff as an escape
value that indicates that we don't know the offset.