Subject: Add preliminary TCP socket support for Windows [s48-tuebingen @ 641]
This requires adding a new file descriptor type, which is why
the code is sitting in fd-io.c, not socket.c.

Note that testing this using "localhost" causes problems because of
some unspeakable Windows bogosity.

I also had to change the signature of s48_accept to require a RETRY?
(to match s48_connect).  This will temporarily break the Unix sockets
(which had a few minor problems anyway) which I'll fix momentarily.