Subject: Get rid of obsolete S48_CFLAG_CKR [s48-tuebingen @ 674]
"Haven't I killed you already?"

The comment referring to IRIX C had ceased to apply to an old version
of IRIX cc in 1996 (when I was at NECI).  I think we can safely assume
all instances of this brokenness are gone now.  (At least where we
might expect the rest of Scheme 48 to build.)