Subject: Support external functions with >12 arguments [s48-tuebingen @ 720]
As posted on s48-devel, this works simply by calling the C function
with two arguments containing the number of Scheme arguments and a
pointer into the stack.  A macro S48_ARG_REF helps translate the stack
indices.

The support in external.c for calling the C function was already
present as part of an earlier patch (more or less inadvertently), but
only now gets activated.