Subject: Fix I/O bug [s48-tuebingen @ 749]
This is a fix for a bug that can only be seen on Windows:

EMPTY-BUFFER!, after sending the buffer contents to the lower levels,
didn't actually wait for the channel condvar, even if it was called
with NECESSARY?.

If nothing was written immediately, this used to mean that automatic
buffer flushing would take care of things.  However, the automatic
buffer flushing checks if there are waiters for a given port, and, if
the I/O completion event hasn't gotten back, abort the I/O by sending
a "0 bytes sent" message.  That, in turn, would lead to the data being
written a second (or third or ...) time.

This didn't matter on Unix as data would only be written synchronously
off CHANNEL-MAYBE-WRITE.

To fix the problem, we add a WAIT? argument to
CHANNEL-MAYBE-COMMIT-AND-WRITE that tells it to wait on the condvar
immediately.