Subject: GC fixes [s48-tuebingen @ 846]
From Richard:

Here is a diff for fixing the two bugs that Jorgen Schaefer
ran into with his SRFI-27+threads code.  A GC when copying the
list during for APPLY left the code vector pointing into the old
heap.  This wasn't noticable unless you happened to get a timer
interrupt and thread switch on that call.

The other bug cause the interrupt flag to be temporarily cleared
if the first post-interrupt call was an APPLY.