Subject: Make the PreScheme type rule for IF more strict [s48-tuebingen @ 865]
A recent post on scheme48-bugs (by Markus Ziegler) showed that the
type inferencer does not work consistently with the C code generation:
The type inferencer allows two branches of an IF to have inconsistent
types, as long as one of them is unit or null---a unit type in one
branch will force the type of the entire IF to unit.  However, the
code generation looks at the values, not the types, and bad things can
happen if they get out of synch.

So this patch changes the type inference rule for IF to enforce that
the types of the two branches of an IF be the same.  This requires a
number of massages to the VM source code---arguably mostly for the
better.