Subject: Add proper implementation of CHAR-READY? for input ports [s48-tuebingen @ 949]
This needs to do text decoding, so it's considerably more complex than
the old implementation.  We do this by splicing yet another option
into the PORT-HANDLER-CHAR procedure.

Note that R5RS CHAR-READY? for output ports is fundamentally broken in
the fact of variable-length character encodings, as we would need to
know the character the program wants to write before knowing that it
can be written without blocking.