Subject: Fix bug in VALUE-PIPES [s48-tuebingen @ 979]
From a report by Taylor Campbell:

BLOCK-ON-PIPE waits on the pipe's queue, and loses if the commit
fails, but just returns #T if the commit succeeds.  Instead, it
should attempt to commit and _unconditionally_ call LOSE.  (It did
this in older versions of Scheme48; this bug is new in 1.1.)  With
the current code, every time a pipe is written to and subsequently
read from, the pipe will produce two values in quick succession: #T &
the desired regular value.  Here is an example that demonstrates the
bug:

      > ,open threads value-pipes formats srfi-31
      > (define pipe (make-pipe))
      ; no values returned
      > (spawn (rec (loop)
                 (format #t "~&Smoked from pipe: ~S~%"
                         (pipe-read! pipe))
                 (loop)))
      #{Unspecific}
      > (pipe-write! pipe 'pot)
      ; no values returned
      > Smoked from pipe: #t
      Smoked from pipe: pot
      (pipe-write! pipe 'weed)
      ; no values returned
      > Smoked from pipe: #t
      Smoked from pipe: weed

  (Jorgen Schaefer noticed the bug; I noticed the fix.)