Subject: Increase precision of EXPT for exact integer exponents [s48-tuebingen @ 1359]
The high-performance code we had was only good for exact integer
bases; for everything else we'd revert to (exp (* n (log x))), which
is less precise than we can do by using divide-and-conquer
multiplications.  (We actually need this to make EXACT->INEXACT work
for the general rational case.  More on that later.)