Subject: Fix bug in OPEN-FILE [s48-tuebingen @ 1422]
This is a continuation of rev 1340, where I changed the logic to
determine whether a channel created by the C code should be an input
or an output channel---sifting around in the Unix option mask is no
good, as any of those mask might be zero.  So we pass an INPUT? flag
in explicitly.

Reported by: Andreas Rottmann