Subject: Fixed SRFI 71 structure definition, passes all tests now. [s48-tuebingen @ 1559]
Fixed the structure definition of SRFI 71 so that all tests included
in the SRFI document now pass.  The fix is to introduce a auxiliary
structure that exports srfi-let/let*/letrec.  The SRFI 71 structure
now opens the auxiliary struct and renames these to let/let*/letrec.
May not be the cleanest solution, but it works.