Subject: Updated floating point conversion code to use exact 2 [s48-tuebingen @ 1560]
The code for converting to/from float point numbers and integers is
based on the reference implementation for SRFI 56, so I've changed 2.0
to 2 as pointed out by Bradley Lucier on the SRFI 56 mailing list.

   http://srfi.schemers.org/srfi-56/post-mail-archive/msg00001.html

Previously this was used only in the 32 bit case; I just missed the 64
case by accident.

Also changed indentation a bit so that u32->fl and u64->fl are
presented consistently.