Subject: Make semantics of DYNAMIC/WIND in/out thunks consistent [s48-tuebingen @ 1599]
In wind.scm, there are two places where the IN and OUT thunks are run:
One in DYNAMIC-WIND itself, and one in TRAVEL-TO-POINT!.  The desired
semantics (the one in DYNAMIC-WIND) is that we've reached a point only
after the IN thunk has run and leave it just before running the OUT
thunk.  TRAVEL-TO-POINT! would reach the point *before* running the IN
thunk, however.  Fix.  (This also makes us adhere to R6RS.)