Subject: Make REMOVE-UNUSED-PROCEDURES! check for empty LET calls [s48-tuebingen @ 1631]
Consider the following program (actually it's the output of a flow
analysis based inlining optimization):

   2   (P olin-inline-3_2 (c_1)
   4     (LET* (((v_3)*     '23)
   14           ((f_13)*    ^f_7)
   21           ((p_20)*    ^p_24))
           (unknown-return 0 ^c_15 p_20)))

   15   (C c_15 (v_16)
          (unknown-return 0 c_1 v_3))

   24   (P p_24 (c_22 y_23)
          (unknown-return 0 c_22 v_3))

   7    (P f_7 (c_6 x_5)
   12     (LET* (((p_11)*    ^p_10))
            (unknown-return 0 c_6 p_11)))

   10    (P p_10 (c_9 y_8)
           (unknown-return 0 c_9 v_3))

In a first step, the byte-code optimizer transforms the program to
this one:

   2   (P olin-inline-3_2 (c_1)
   21    (LET* (((p_20)*    ^p_24))
           (unknown-return 0 ^c_15 p_20)))

   15   (C c_15 (v_16)
          (unknown-return 0 c_1 '23))

   24   (P p_24 (c_22 y_23)
          (unknown-return 0 c_22 '23))

Now, the optimizer calls INTEGRATE-JUMP-PROCS! and removes the unused
procedures (see REMOVE-UNUSED-PROCEDURES! in
ps-compiler/front/jump.scm).  In this case REMOVE-UNUSED-PROCEDURES!
removes all call arguments for the LET call (i.e. the body of
continuation lambda 21).  The result looks like this:

  2   (P olin-inline-3_2 (c_1 e_28 t_29)
  15    (LET* ((()))
          (unknown-return 0 c_1 '23)))

The byte-code generator for LET in bc-optimizer/bc-base.scm, however,
assumes that LET calls have at least one argument (besides the
continuation) and fails.

This patch makes REMOVE-UNUSED-PROCEDURES! check whether it just
removed the last non-continuation argument of a LET call.  If this is
the case, it replaces the LET call with the body of the continuation.