Subject: Add OP-WITH-CELL-LITERAL instruction [s48-tuebingen @ 1710]
This instruction allows embedding of a cell-wide literal along with an opcode in the instruction stream to minimize pushing of arguments
for the benefit of the native-code compiler.

It takes as its argument the literal and an opcode and passes the literal as the first argument to the opcode. *val* contains the
second argument and the remaining arguments are on the stack. The VM implements this instruction by pushing *val*, loading the
literal into *val*, and jumping to the offset of the opcode. Thus, from the point of view of the VM, other instruction-stream
arguments could follow the opcode. This would however break the tools operating on the byte-code and properly supporting this would
require a new argument type that can encompass an opcode and its instruction-stream arguments.

The cell is currently represented as a sequence of 4 bytes in the definition of the instruction set but of course we should introduce a
new argument type for cells instead.