Subject: Add configure option --enable-universal-binary [s48-tuebingen @ 1712]
This patch adds the configure option --enable-universal-binary.  This option
adds the Darwin/OS X specific compiler flags "-arch i386 -arch ppc" to the
CFLAGS to build a so-called "Universal Binary" which contains binaries for both
OS X platforms.  This option also changes the check for endianess:  If
--enable-universal-binary is given, check using the (magic) __BIG_ENDIAN__
macro pre-defined by the compiler.