Subject: Add HANDLE-NATIVE-POLL and RESUME-NATIVE-POLL  [s48-tuebingen @ 1788]
This adds a new instruction RESUME-NATIVE-POLL to be used in
*native-poll-return-code* for jumping back to the native code after
HANDLE-NATIVE-POLL has setup a corresponding adlib continuation. The
post native dispatcher calls HANDLE-NATIVE-POLL, assuming the native
code to have pushed the resuming address (word aligned, so it does not
hurt the GC) and the template, and to have loaded
*VAL*. RESUME-NATIVE-POLL uses a new glue code function
S48-JUMP-NATIVE that can directly jump to a native code address.

The good thing about this is that it requires very little work (and
code) on the side of the native code. However, passing the resuming
address directly is slightly ugly.

It would be nice to combine this with the interrupt handling code for
native calls but this requires some substantial rewriting of very
low-level code.