Subject: use sigaltstack() if native code is enabled [s48-tuebingen @ 2009]
Usually the signal handler runs on the same stack as the program. This
may lead to strange bug if the stack pointer is temporarily set below
the real top of the stack and a signal handler runs. As the native
code is directly manipulating the stack pointer such an inconsistent
stack may occur. Allocating extra stack space for the signal handlers
frees the native code from ensuring that the stack pointer always
points to the top of the stack.