Make invalid-assignment warning more consistent.
As only one thingie for a variable goes into a template, the
information that the variable is assigned may get lost.  The following
structure definitions don't cause a warning when BAZ is loaded, as
they should:

(define-structure foo (export bar)
  (open scheme)
  (begin
    (define bar 'foo)))

(define-structure baz (export)
  (open scheme foo)
  (begin
    (lambda ()
      bar
      (set! bar 'bla))))

(Note that the warning appears when the first reference to BAR is
removed.)

To remedy this, reduce the type information in THINGIE to "assigned or
not" (the compiler only ever used VALUE-TYPE and USUAL-VARIABLE-TYPE
anyway), and update it when re-using an existing thingie.