Fixed wrong area lookups of stobs in bibop, resulting in segfaults.
There can be empty objects (e.g. vectors of size 0), so we always have to take the
header's address to look up the area. The address after the header might point to a
different area or even outside the scheme heap, if the object is placed at the very end
of an area.